Implement ssf.Notification API with Promise based permission
Also removed the `permission` getter, since it isn't part of the TypeScript
definition on the Window object. For client code it's much cleaner to always
call the promise function rather than check the property to decide whether
to call the promise.
