+++ packages/api-browser/index.ts
@@ -2,10 +2,12 @@ import app from './src/app';
 import MessageService from './src/message-service';
 import ScreenSnippet from './src/screen-snippet';
 import Window from './src/window';
+import Notification from './src/notification';
 
 export {
   app,
   MessageService,
   ScreenSnippet,
-  Window
+  Window,
+  Notification
 };
