+++ packages/api-demo/src/notifications/notification-demo.js
@@ -1,29 +1,9 @@
 var button = document.getElementById('notification-test');
 
-// if the environment this code is running in requires that the user grants permission
-// before raising a notification, then do so!
-function withNotificationPermission() {
-  return new Promise(function(resolve, reject) {
-    // if permission already granted, we're good to go
-    Notification.requestPermission(function(permission) {
-      if (permission === 'granted') {
-        resolve(true);
-      } else {
-        resolve(false);
-      }
-    });
-  });
-}
-
 button.onclick = function() {
-  if (!('Notification' in window)) {
-    console.error('This browser / container does not support system notifications');
-    return;
-  }
-
-  withNotificationPermission()
-    .then(function(hasPermission) {
-      if (!hasPermission) {
+  ssf.Notification.requestPermission()
+    .then(function(permission) {
+      if (permission !== 'granted') {
         console.error('notification permission not granted');
         return;
       }
@@ -31,7 +11,7 @@ button.onclick = function() {
       var title = document.getElementById('title').value;
       var body = document.getElementById('body').value;
       // eslint-disable-next-line no-new
-      new Notification(title, {
+      new ssf.Notification(title, {
         body: body
       });
     });
