+++ packages/api-electron/preload.ts
@@ -1,12 +1,13 @@
-import './src/preload/shims';
 import app from './src/preload/app';
 import MessageService from './src/preload/message-service';
 import ScreenSnippet from './src/preload/screen-snippet';
 import Window from './src/preload/window';
+import Notification from './src/preload/notification';
 
 export {
   app,
   MessageService,
   ScreenSnippet,
-  Window
+  Window,
+  Notification
 };
