+++ packages/api-openfin/index.ts
@@ -1,12 +1,13 @@
-import './src/shims';
 import app from './src/app';
 import MessageService from './src/message-service';
 import ScreenSnippet from './src/screen-snippet';
 import Window from './src/window';
+import Notification from './src/notification';
 
 export {
   app,
   MessageService,
   ScreenSnippet,
-  Window
+  Window,
+  Notification
 };
